backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
google-generativeai==0.7.2
